Fix possible infinite loop in eh_frame_hdr parsing
If cie_fde_size is equal to (ULONG_MAX - (4 + 1)), then the variable ptr
overflows to its original value. In this case, the same cie_fde_size is
read again, and the loop never terminates.

Keep the previous pointer value and ensure that the ptr increases. If the
pointer stagnates or goes backward, then terminate the loop.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>